+++ master/buildbot/test/fake/fakedb.py
@@ -769,10 +769,7 @@ class FakeBuildsetsComponent(FakeDBComponent):
 
         # only add brids if we're expecting them (sometimes they're unknown)
         if 'brids' in expected_buildset:
-            brids = dict([ (br.buildername, br.id)
-                      for br in self.db.buildrequests.reqs.values()
-                      if br.buildsetid == bsid ])
-            buildset['brids'] = brids
+            buildset['brids'] = self.allBuildRequests(bsid)
 
         if 'id' in ss:
             del ss['id']
@@ -796,6 +793,15 @@ class FakeBuildsetsComponent(FakeDBComponent):
         self.t.assertEqual(sorted(subscriptions),
                          sorted(self.buildset_subs))
 
+    def allBuildRequests(self, bsid=None):
+        if bsid is not None:
+            is_same_bsid = lambda br: br.buildsetid==bsid
+        else:
+            is_same_bsid = lambda br: True
+        return dict([ (br.buildername, br.id)
+              for br in self.db.buildrequests.reqs.values()
+              if is_same_bsid(br) ])
+        
 
 class FakeStateComponent(FakeDBComponent):