+++ master/docs/release-notes.rst
@@ -81,6 +81,11 @@ Features
   argument.  This allows substitution of multiple command-line arguments using
   properties.  See :bb:bug:`2150`.
 
+* :py:class:`Trigger` steps with ``waitForFinish=True`` now receive a URL to
+  all the triggered builds. This URL is displayed in the waterfall and build
+  details. Also, the statistic ``triggered_builds`` is set in the ``StepStatus``
+  to a list of build dictionaries. See :bb:bug:`2170`.
+
 Slave
 -----